+++ rust/src/config/mod.rs
@@ -2,8 +2,10 @@ mod scaling_quality;
 mod vanilla_version;
 mod resolution;
 mod ja2_json;
+mod cli;
 
 pub use self::scaling_quality::ScalingQuality;
 pub use self::vanilla_version::VanillaVersion;
 pub use self::resolution::Resolution;
 pub use self::ja2_json::Ja2Json;
+pub use self::cli::Cli;
