+++ rust/src/stracciatella.rs
@@ -9,30 +9,20 @@ extern crate dirs;
 
 use std::slice;
 use std::str;
-use std::str::FromStr;
 use std::ptr;
-use std::fs;
 use std::ffi::{CStr, CString};
 use std::path::PathBuf;
 use std::default::Default;
-use serde::Deserialize;
-use serde::Serialize;
 
-use getopts::Options;
 use libc::{size_t, c_char};
 
 mod config;
 
 pub use config::ScalingQuality;
 pub use config::VanillaVersion;
-pub use config::Resolution;
-pub use config::Ja2Json;
-
-#[cfg(not(windows))]
-static DATA_DIR_OPTION_EXAMPLE: &'static str = "/opt/ja2";
-
-#[cfg(windows)]
-static DATA_DIR_OPTION_EXAMPLE: &'static str = "C:\\JA2";
+use config::Resolution;
+use config::Ja2Json;
+use config::Cli;
 
 #[derive(Debug, PartialEq)]
 pub struct EngineOptions {
@@ -73,168 +63,9 @@ impl Default for EngineOptions {
     }
 }
 
-pub fn get_command_line_options() -> Options {
-    let mut opts = Options::new();
-
-    opts.long_only(true);
-
-    opts.optmulti(
-        "",
-        "datadir",
-        "Set path for data directory",
-        DATA_DIR_OPTION_EXAMPLE
-    );
-    opts.optmulti(
-        "",
-        "mod",
-        "Start one of the game modifications. MOD_NAME is the name of modification, e.g. 'from-russia-with-love. See mods folder for possible options'.",
-        "MOD_NAME"
-    );
-    opts.optopt(
-        "",
-        "res",
-        "Screen resolution, e.g. 800x600. Default value is 640x480",
-        "WIDTHxHEIGHT"
-    );
-    opts.optopt(
-        "",
-        "brightness",
-        "Screen brightness (gamma multiplier) value to set where 0.0 is completely dark and 1.0 is normal brightness. Default value is 1.0",
-        "GAMMA_VALUE"
-    );
-    opts.optopt(
-        "",
-        "resversion",
-        "Version of the game resources. Possible values: DUTCH, ENGLISH, FRENCH, GERMAN, ITALIAN, POLISH, RUSSIAN, RUSSIAN_GOLD. Default value is ENGLISH. RUSSIAN is for BUKA Agonia Vlasty release. RUSSIAN_GOLD is for Gold release",
-        "RUSSIAN_GOLD"
-    );
-    opts.optflag(
-        "",
-        "unittests",
-        "Perform unit tests. E.g. 'ja2.exe -unittests --gtest_output=\"xml:report.xml\" --gtest_repeat=2'");
-    opts.optflag(
-        "",
-        "editor",
-        "Start the map editor (Editor.slf is required)"
-    );
-    opts.optflag(
-        "",
-        "fullscreen",
-        "Start the game in the fullscreen mode"
-    );
-    opts.optflag(
-        "",
-        "nosound",
-        "Turn the sound and music off"
-    );
-    opts.optflag(
-        "",
-        "window",
-        "Start the game in a window"
-    );
-    opts.optflag(
-        "",
-        "debug",
-        "Enable Debug Mode"
-    );
-    opts.optflag(
-        "",
-        "help",
-        "print this help menu"
-    );
-
-    opts
-}
-
 fn parse_args(engine_options: &mut EngineOptions, args: &[String]) -> Option<String> {
-    let opts = get_command_line_options();
-
-    match opts.parse(&args[1..]) {
-        Ok(m) => {
-            if !m.free.is_empty() {
-                return Some(format!("Unknown arguments: '{}'.", m.free.join(" ")));
-            }
-
-            if let Some(s) = m.opt_str("datadir") {
-                match fs::canonicalize(PathBuf::from(s)) {
-                    Ok(s) => {
-                        let mut temp = String::from(s.to_str().expect("Should not happen"));
-                        // remove UNC path prefix (Windows)
-                        if temp.starts_with("\\\\") {
-                            temp.drain(..2);
-                            let pos = temp.find('\\').unwrap() + 1;
-                            temp.drain(..pos);
-                        }
-                        engine_options.vanilla_data_dir = PathBuf::from(temp)
-                    },
-                    Err(_) => return Some(String::from("Please specify an existing datadir."))
-                };
-            }
-
-            if !m.opt_strs("mod").is_empty() {
-                engine_options.mods = m.opt_strs("mod");
-            }
-
-            if let Some(s) = m.opt_str("res") {
-                match Resolution::from_str(&s) {
-                    Ok(res) => {
-                        engine_options.resolution = res;
-                    },
-                    Err(s) => return Some(s)
-                }
-            }
-
-            if let Some(s) = m.opt_str("brightness") {
-                match s.parse::<f32>() {
-                    Ok(val) => {
-                        engine_options.brightness = val;
-                    },
-                    Err(_e) => return Some(String::from("Incorrect brightness value."))
-                }
-            }
-
-            if let Some(s) = m.opt_str("resversion") {
-                match VanillaVersion::from_str(&s) {
-                    Ok(resource_version) => {
-                        engine_options.resource_version = resource_version
-                    },
-                    Err(str) => return Some(str)
-                }
-            }
-
-            if m.opt_present("help") {
-                engine_options.show_help = true;
-            }
-
-
-            if m.opt_present("unittests") {
-                engine_options.run_unittests = true;
-            }
-
-            if m.opt_present("editor") {
-                engine_options.run_editor = true;
-            }
-
-            if m.opt_present("fullscreen") {
-                engine_options.start_in_fullscreen = true;
-            }
-
-            if m.opt_present("nosound") {
-                engine_options.start_without_sound = true;
-            }
-
-            if m.opt_present("window") {
-                engine_options.start_in_window = true;
-            }
-
-            if m.opt_present("debug") {
-                engine_options.start_in_debug_mode = true;
-            }
-
-            None
-        }
-        Err(f) => Some(f.to_string())
-    }
+    let cli = Cli::from_args(args);
+    cli.apply_to_engine_options(engine_options).err()
 }
 
 pub fn ensure_json_config_existence(stracciatella_home: &PathBuf) -> Result<(), String> {
@@ -322,9 +153,7 @@ pub extern fn create_engine_options(array: *const *const c_char, length: size_t)
     match build_engine_options_from_env_and_args(&args) {
         Ok(engine_options) => {
             if engine_options.show_help {
-                let opts = get_command_line_options();
-                let brief = "Usage: ja2 [options]".to_string();
-                print!("{}", opts.usage(&brief));
+                print!("{}", Cli::usage());
             }
             Box::into_raw(Box::new(engine_options))
         },
